batch-daily-sp500ma-notification/requirements.txt
requests
boto3
botocore
python-dotenv
